datapath query (ref, follow-up)
