+++ ais/tgts3.go
@@ -150,7 +150,7 @@ func (t *target) directPutObjS3(w http.ResponseWriter, r *http.Request, items []
 
 	dpq := dpqAlloc()
 	defer dpqFree(dpq)
-	if err := urlQuery(r.URL.RawQuery, dpq); err != nil {
+	if err := dpq.fromRawQ(r.URL.RawQuery); err != nil {
 		t.writeErr(w, r, err)
 		return
 	}
@@ -193,7 +193,7 @@ func (t *target) getObjS3(w http.ResponseWriter, r *http.Request, items []string
 		return
 	}
 	dpq := dpqAlloc()
-	if err := urlQuery(r.URL.RawQuery, dpq); err != nil {
+	if err := dpq.fromRawQ(r.URL.RawQuery); err != nil {
 		dpqFree(dpq)
 		t.writeErr(w, r, err)
 		return
